+++ src/main/java/blusunrize/immersiveengineering/common/items/RobotWolfItem.java
@@ -8,12 +8,16 @@
 
 package blusunrize.immersiveengineering.common.items;
 
+import blusunrize.immersiveengineering.api.IETags;
+import blusunrize.immersiveengineering.api.Lib;
+import blusunrize.immersiveengineering.api.utils.ItemUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.sounds.SoundEvents;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobSpawnType;
@@ -21,17 +25,23 @@ import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.animal.Wolf;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.SpawnEggItem;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.gameevent.GameEvent;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.fml.common.EventBusSubscriber.Bus;
+import net.neoforged.neoforge.event.entity.player.PlayerInteractEvent.EntityInteractSpecific;
 
 import java.util.List;
 import java.util.Objects;
 
 import static blusunrize.immersiveengineering.api.IEApi.ieLoc;
 
+@EventBusSubscriber(modid = Lib.MODID, bus = Bus.GAME)
 public class RobotWolfItem extends IEBaseItem
 {
 	public static ResourceLocation REGISTRY_KEY = ieLoc("robot");
@@ -84,6 +94,34 @@ public class RobotWolfItem extends IEBaseItem
 	@Override
 	public void appendHoverText(ItemStack stack, TooltipContext ctx, List<Component> list, TooltipFlag flag)
 	{
+	}
 
+	@SubscribeEvent
+	public static void onWolfInteract(EntityInteractSpecific event)
+	{
+		if(!(event.getTarget() instanceof Wolf wolf)||!wolf.getVariant().is(RobotWolfItem.REGISTRY_KEY))
+			// Only affect robot wolves
+			return;
+
+		ItemStack stack = event.getItemStack();
+		if(wolf.isFood(stack)||stack.getItem() instanceof SpawnEggItem)
+		{
+			// Robot wolves do not use normal food and don't interact with spawn eggs
+			event.setCanceled(true);
+			event.setCancellationResult(InteractionResult.sidedSuccess(event.getLevel().isClientSide()));
+		}
+		else if(wolf.getHealth() < wolf.getMaxHealth())
+		{
+			// Robot wolves are repaired with hammers & plates
+			ItemUtils.isHoldingBoth(event.getEntity(), IETags.hammers, IETags.plates).ifPresent(heldItems -> {
+				wolf.heal(16);
+				// Damage the hammer, consume the plate
+				heldItems.getFirst().stack().hurtAndBreak(1, event.getEntity(), heldItems.getFirst().slot());
+				heldItems.getSecond().stack().consume(1, event.getEntity());
+				event.getEntity().playSound(SoundEvents.ANVIL_USE);
+				event.setCanceled(true);
+				event.setCancellationResult(InteractionResult.sidedSuccess(event.getLevel().isClientSide()));
+			});
+		}
 	}
 }